refactor(foundation): move GOOSESelect event to plugins foundation
